pyproject.toml
[project]
name = "scrap-cityevent"
version = "0.1.0"
description = "A script to monitor events containing 'Atelier' and send notifications"
authors = [
    {name = "Antonia"}
]
readme = "README.md"
requires-python = "^3.8"
dependencies = [
    "requests (>=2.32.3,<3.0.0)",
    "beautifulsoup4 (>=4.13.3,<5.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)"
]

[project.scripts]
scrap-cityevent = "scrap_cityevent.main:main"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
